[UPDATE] 스타일 분리작업
